raspi-streamer/requirements.txt
absl-py==0.7.0
astroid==1.6.5
certifi==2018.4.16
chardet==3.0.4
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
linecache2==1.0.0
mccabe==0.6.1
nose==1.3.7
prometheus-client==0.6.0
psutil==5.4.5
pylint==1.9.2
PyOpenGL==3.1.5
PyYAML==3.13
rectpack==0.2.1
requests==2.18.4
six==1.15.0
sseclient==0.0.22
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.22
wrapt==1.10.11
